Move most geopoint map logic to Fragment
