+++ geo/src/main/java/org/odk/collect/geo/DaggerSetup.kt
@@ -11,6 +11,7 @@ import org.odk.collect.async.Scheduler
 import org.odk.collect.geo.geopoint.GeoPointActivity
 import org.odk.collect.geo.geopoint.GeoPointDialogFragment
 import org.odk.collect.geo.geopoint.GeoPointMapActivity
+import org.odk.collect.geo.geopoint.GeoPointMapFragment
 import org.odk.collect.geo.geopoint.GeoPointViewModelFactory
 import org.odk.collect.geo.geopoint.LocationTrackerGeoPointViewModel
 import org.odk.collect.geo.geopoly.GeoPolyFragment
@@ -49,6 +50,7 @@ interface GeoDependencyComponent {
     fun inject(geoPointActivity: GeoPointActivity)
     fun inject(selectionMapFragment: SelectionMapFragment)
     fun inject(geoPolyFragment: GeoPolyFragment)
+    fun inject(geoPointMapFragment: GeoPointMapFragment)
 
     val scheduler: Scheduler
     val locationTracker: LocationTracker
